creacion de modelso y ajuste para la carga de excel e insert masivo

+++ database/migrations/2024_01_10_133759_create_master_lpa_personas_table.php
@@ -79,7 +79,7 @@ class CreateMasterLpaPersonasTable extends Migration
             $table->string('discapacidad_comunicar')->nullable();
             
             //Celular	(numero)
-            $table->tinyInteger('telefono')->nullable();
+            $table->bigInteger('telefono')->nullable();
             
             $table->timestamps();
         });
